requirements.txt
flask
flask-cors
flask-socketio
python-dotenv
supabase
pyjwt
requests
gunicorn
gevent
gevent-websocket